# DDS: optimize loading of compressed images (3-5x)