@aras-p 2022-10-05
Huh, the two CI failures are the most curious. I don't yet quite see how they would be related to the single .cpp source file change here.

The `Ubuntu bleeding edge gcc12 C++20 py3.10 OCIO/libtiff/exr-master boost1.74 avx2` one fails linking `lib/libOpenImageIO.so.2.5.0` with an error coming from libtiff (?):
```
/usr/bin/ld: /home/runner/work/oiio/oiio/ext/dist/lib/libtiff.a(tif_unix.c.o): warning: relocation against `stderr@@GLIBC_2.2.5' in read-only section `.text'
/usr/bin/ld: /home/runner/work/oiio/oiio/ext/dist/lib/libtiff.a(tif_unix.c.o): relocation R_X86_64_PC32 against symbol `stderr@@GLIBC_2.2.5' can not be used when making a shared object; recompile with -fPIC
```

The failure on `windows-2019 VS2019` is where `unit_timer` tests fail:
```
141/157 Test #144: unit_timer ..............................***Failed    2.90 sec
Timer begin/end cost is 2.13647e+07 /sec
Out of 10000000 queries, 7785692 had no time
Longest was 0.0007168 s.
Checkpoint: All 0.438635 selective 0.150176
D:/a/oiio/oiio/src/libutil/timer_test.cpp:105:
FAILED: all() == 0.5
	values were '0.753666' and '0.5', diff was 0.253656
D:/a/oiio/oiio/src/libutil/timer_test.cpp:112:
FAILED: all() == 0.5
	values were '0.753693' and '0.5', diff was 0.253692
D:/a/oiio/oiio/src/libutil/timer_test.cpp:116:
FAILED: all() == 0.6
	values were '0.897358' and '0.6', diff was 0.297346
Checkpoint2: All 0.897369 selective 0.301812
string ctr      :    0.0 ns (+/- 0.0ns), unreliable
usleep(1000)    :   40.4 ms (+/-28.2ms),    0.0 k/s
add             :    1.6 ns (+/- 0.0ns),  624.6 M/s
cos             :    5.6 ns (+/- 0.0ns),  178.1 M/s
acos            :    8.5 ns (+/- 0.2ns),  117.1 M/s
fast_acos       :    3.2 ns (+/- 0.0ns),  311.9 M/s
sqrt            :    1.6 ns (+/- 0.0ns),  623.7 M/s
simd sqrt       :    1.2 ns (+/- 0.0ns), 3325.0 Mvals/s, 831.2 Mcalls/s
ERRORS!
```