@lgritz 2022-10-06
Aside for @aras-p to know next time this comes up: a good way to time how long it takes to do a full read with minimal conversions and no SHA-1 overhead is

    time oiiotool --native -i:now=1 file.exr

The `now=1` modifier causes the input to be read in full rather than going through the ImageCache, and all at once. And the `--native` keeps it in the data type of the file itself if at all possible (versus the oiiotool default of tending to convert everything to float upon input so that image processing operations are done at full speed and precision).